refactor: ChartMetaBuilder を OcPageCacheDataBuilder にリネーム（キャッシュ用データ抽出ハブへ）
「チャートの可用性メタ専用(ChartMeta)」という体裁から、per-room で複数のキャッシュ用
データ(チャート可用性メタ＋rising掲載状態＋今後の相乗り)を抽出・組み立てるハブ、という
役割に合わせて改称した。perf都合で per-room 取得に密結合せざるを得ないため、同種の派生
データの相乗りは今後もこのクラスに集約する想定。

- App\Services\Statistics\ChartMeta\ChartMetaBuilder
  → App\Services\StaticData\OcPageCacheDataBuilder（保存役の OcPageCacheGenerator と対）
- テストも app/Services/StaticData/test/OcPageCacheDataBuilderTest.php へ移動
- 注入変数 $chartMetaBuilder → $cacheDataBuilder、同一namespaceの冗長useを除去
- ChartAvailabilityCalculator は chart 固有のため現状維持（cross-namespace use を追加）
- 挙動は不変（リネームのみ）。phpunit 24件・autoload 再生成で確認

Committed from: user-B550M-Pro4:~/repos/oc-graph-mock
